refactor(agents): split embedded runner orchestration
