+++ src/agents/embedded-agent-runner/run/attempt-startup.test.ts
@@ -1,0 +1,61 @@
+import { beforeEach, describe, expect, it, vi } from "vitest";
+import type { EmbeddedRunAttemptParams } from "./types.js";
+
+const mocks = vi.hoisted(() => ({
+  applySkillEnvOverrides: vi.fn(),
+  mapSandboxSkillEntriesForPrompt: vi.fn(),
+}));
+
+vi.mock("../../../skills/runtime/env-overrides.js", () => ({
+  applySkillEnvOverrides: mocks.applySkillEnvOverrides,
+  applySkillEnvOverridesFromSnapshot: vi.fn(),
+}));
+
+vi.mock("../../../skills/runtime/embedded-run-entries.js", () => ({
+  resolveEmbeddedRunSkillEntries: vi.fn(() => ({
+    shouldLoadSkillEntries: true,
+    skillEntries: [],
+  })),
+}));
+
+vi.mock("../../../skills/loading/workspace.js", () => ({
+  resolveSkillsPromptForRun: vi.fn(() => "skills prompt"),
+}));
+
+vi.mock("../sandbox-skills.js", () => ({
+  resolveSandboxSkillRuntimeInputs: vi.fn(() => ({
+    skillsEligibility: undefined,
+    skillsPromptWorkspaceDir: "/tmp/workspace",
+    skillsSnapshot: undefined,
+    skillsWorkspaceDir: "/tmp/workspace",
+    workspaceOnly: false,
+  })),
+  mapSandboxSkillEntriesForPrompt: mocks.mapSandboxSkillEntriesForPrompt,
+  mapSandboxSkillUsagePaths: vi.fn(() => []),
+}));
+
+import { prepareEmbeddedAttemptSkills } from "./attempt-startup.js";
+
+describe("prepareEmbeddedAttemptSkills", () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
+
+  it("restores environment overrides when later preparation fails", () => {
+    const restore = vi.fn();
+    mocks.applySkillEnvOverrides.mockReturnValue(restore);
+    mocks.mapSandboxSkillEntriesForPrompt.mockImplementation(() => {
+      throw new Error("skill prompt mapping failed");
+    });
+
+    expect(() =>
+      prepareEmbeddedAttemptSkills({
+        attempt: { config: {} } as EmbeddedRunAttemptParams,
+        effectiveWorkspace: "/tmp/workspace",
+        sandbox: undefined,
+        sessionAgentId: "main",
+      }),
+    ).toThrow("skill prompt mapping failed");
+    expect(restore).toHaveBeenCalledOnce();
+  });
+});
